action, reducer file name change and code refactor and introduce new helper
