--- src/services/actions/addTodo.js
@@ -1,10 +1,0 @@
-export const addTodo = (props) => {
-  const { text, createdAt } = props;
-  return {
-    type: "ADD",
-    payload: {
-      text: text,
-      createdAt: createdAt,
-    },
-  };
-};
