+++ src/services/actions/todo.js
@@ -1,0 +1,6 @@
+export const addTodo = (props) => {
+    return {
+        type: "ADD",
+        payload: props,
+    };
+};
